Add a "shouldUpdate" method which can be easily edited.
Because of all the confusion regarding different versioning schemes and trying to use them to determine the "validity" of a file present on BukkitDev as a "new" version, Updater only does a string comparison to see if the newest file's version is different than the current version. If so, it will consider it a "valid" update.

I've added a new method that will allow anyone who has a specific, constant versioning scheme that they use in their plugins to more easily implement a mathematical comparison.